+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/video/capture/CaptureSourceError.kt
@@ -16,6 +16,11 @@ enum class CaptureSourceError {
     Unknown,
 
     /**
+     * A failure to obtain necessary permission to start video
+     */
+    PermissionError,
+
+    /**
      * A failure observed from a system API used for capturing
      * e.g. In response to a `CameraDevice.StateCallback().onError` call
      */