+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/session/MeetingSession.kt
@@ -5,6 +5,7 @@
 
 package com.amazonaws.services.chime.sdk.meetings.session
 
+import com.amazonaws.services.chime.sdk.meetings.analytics.EventAnalyticsController
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.AudioVideoFacade
 import com.amazonaws.services.chime.sdk.meetings.utils.logger.Logger
 
@@ -16,4 +17,5 @@ interface MeetingSession {
     val configuration: MeetingSessionConfiguration
     val logger: Logger
     val audioVideo: AudioVideoFacade
+    val eventAnalyticsController: EventAnalyticsController
 }